Update MAC_SDK from 4.34 to 4.44
Version 4.44 of MAC_SDK is a version, which is still available for
download and contains the changes to `Seek()` in IO.h, which have
been introduced in 4.41 and 4.42 [1].

- `APE::CIO::Seek()` is split now into `SetSeekPosition()`,
  `SetSeekMethod()` and `PerformSeek()`
- Use the introduced `m_nSeekPosition` and `m_nSeekMethod` variables
  in the `CIO_Seek()` callback function
- `m_nSeekPosition` is `APE::int64` now, whereas the previously used
  `nDistance` was of type `intn`. Use `long` in the callback function
  accordingly.
- Update ThirdParty_MAC_SDK_CUETools.patch for 4.44
- Original download location:
  http://monkeysaudio.com/files/MAC_SDK_444.zip
  SHA256:
  e2af2e9d57b7cd66e5d6525fcd8fcde3839f48a623d31aca22a92cb0d0ea297c

[1] https://monkeysaudio.com/versionhistory.html

+++ CUETools.Codecs.MACLib/MACLib.cs
@@ -57,7 +57,7 @@ namespace CUETools.Codecs.MACLib
     [UnmanagedFunctionPointer(CallingConvention.StdCall)]
     internal unsafe delegate uint CIO_GetSizeDelegate(void* pUserData);
     [UnmanagedFunctionPointer(CallingConvention.StdCall)]
-    internal unsafe delegate int CIO_SeekDelegate(void* pUserData, IntPtr delta, int mode);
+    internal unsafe delegate long CIO_SeekDelegate(void* pUserData, long delta, int mode);
 
     [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Ansi, Pack = 2)]
     public unsafe struct WAVEFORMATEX
